Corrected columns.

+++ SimpleTwitterApp/SimpleTwitterApp.API/Entities/TweetEntity.cs
@@ -4,15 +4,15 @@ namespace SimpleTwitterApp.API.Entities
 {
     public class TweetEntity
     {
-        public int TweetId { get; set; }
+        public long TweetId { get; set; }
 
         public string TweetContent { get; set; }
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
-        public int UserDeviceId { get; set; }
+        public long UserDeviceId { get; set; }
 
-        public int TwittedBy { get; set; }
+        public long TwittedBy { get; set; }
 
         public DateTime TweetTime { get; set; }
     }
